+++ codex/views/opds/v2/const.py
@@ -7,6 +7,7 @@ from types import MappingProxyType
 
 from django.db.models.query import QuerySet
 
+from codex.views.const import BROWSER_FILTER_KEYS
 from codex.views.opds.const import BookmarkFilters, Rel
 
 
@@ -73,10 +74,31 @@ TOP_GROUPS = (
         ),
     ),
 )
+
+_START_QUERY_PARAMS = MappingProxyType(
+    {
+        "topGroup": "p",
+        "filters": {
+            **BookmarkFilters.NONE,
+            **dict.fromkeys(BROWSER_FILTER_KEYS, ()),
+        },
+        "orderBy": "sort_name",
+        "orderReverse": False,
+    }
+)
+
 START_GROUPS = (
     LinkGroup(
-        "Start",
-        (Link(Rel.START, "Start", None, {}, inherit_query_params=False),),
+        "Start (Reset filters & order)",
+        (
+            Link(
+                Rel.START,
+                "Start",
+                None,
+                _START_QUERY_PARAMS,
+                inherit_query_params=False,
+            ),
+        ),
     ),
 )
 FACETS = (